% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTMM.R
\name{MTMM}
\alias{MTMM}
\title{Multivariate Text Mining Module}
\usage{
MTMM()
}
\value{
MTMM is a graphic interface
}
\description{
Multivariante Text Mining Module (MTMM) is a user – oriented interface that integrates the functions from the exploratory analysis method in a text. This interface covers descriptive statistics that frequency tables and plot (Word cloud and Zipf's law), Analysis of similarity, Correspondence Factorial Analysis (AFC) and Reinert.
You can learn more about this package at:
http://www.uv.mx/personal/nehuerta/mtmm/
}
\details{
MTMM is a package with a graphical user interface dedicated to perform of Multivariate Text Mining Analysis in R.

Note: MTMM is free software and comes with ABSOLUTELY NO WARRANTY.
}
\examples{
\dontrun{
##Install package
library(MTMM)
##Call the package
MTMM()
}

}
\author{
Luis Jesus Lozano-Duran, Nery Sofia Huerta-Pacheco, Jose Guillermo Hernandez-Calderon, Sergio Hernandez-Gonzalez
}
